refactor(maya): extract MayaChainAPI TargetType to its own file
